fix: Daemonization for X client
The Xclient needs to be daemonized in order to persistently use the
values set in the clipboard. This is a similar technique used in other
clipboard applications such as xsel and xclip.

The daemon exits when it gets a SelectionClear event, i.e. when Xclient
loses ownership of the selection.
https://tronche.com/gui/x/xlib/events/client-communication/selection-clear.html